Fix Next.js build and clean up
