--- src/pages/main-page.xml
@@ -1,85 +1,0 @@
-<?xml version="1.0" encoding="utf-8" ?>
-<Page xmlns="http://schemas.nativescript.org/tns.xsd" navigatingTo="onNavigatingTo">
-    <TabView selectedIndex="{{ selectedTab }}" androidTabsPosition="bottom">
-        <!-- Home Tab -->
-        <TabViewItem title="Home" iconSource="res://home">
-            <GridLayout>
-                <ScrollView>
-                    <StackLayout>
-                        <!-- Nearby Users -->
-                        <Label text="Nearby People" class="h2 m-10"/>
-                        <ScrollView orientation="horizontal">
-                            <StackLayout orientation="horizontal">
-                                <Repeater items="{{ nearbyUsers }}">
-                                    <Repeater.itemTemplate>
-                                        <StackLayout class="user-card m-5">
-                                            <Image src="{{ dp_url }}" class="profile-image"/>
-                                            <Label text="{{ name }}" class="h3"/>
-                                            <Label text="{{ bio }}" class="body"/>
-                                        </StackLayout>
-                                    </Repeater.itemTemplate>
-                                </Repeater>
-                            </StackLayout>
-                        </ScrollView>
-
-                        <!-- Posts Feed -->
-                        <Label text="Posts & Reels" class="h2 m-10"/>
-                        <Repeater items="{{ posts }}">
-                            <Repeater.itemTemplate>
-                                <StackLayout class="post-card m-10">
-                                    <Image src="{{ media_url }}" class="post-image"/>
-                                    <Label text="{{ title }}" class="h3"/>
-                                    <Label text="{{ caption }}" class="body"/>
-                                </StackLayout>
-                            </Repeater.itemTemplate>
-                        </Repeater>
-                    </StackLayout>
-                </ScrollView>
-            </GridLayout>
-        </TabViewItem>
-
-        <!-- Radar Tab -->
-        <TabViewItem title="Radar" iconSource="res://radar">
-            <ScrollView>
-                <StackLayout>
-                    <Repeater items="{{ nearbyUsers }}">
-                        <Repeater.itemTemplate>
-                            <GridLayout columns="auto, *, auto" class="radar-card m-10">
-                                <Image src="{{ dp_url }}" class="profile-image-small" col="0"/>
-                                <StackLayout col="1">
-                                    <Label text="{{ name }}" class="h3"/>
-                                    <Label text="{{ bio }}" class="body"/>
-                                </StackLayout>
-                                <Button text="Message" class="btn btn-primary" col="2" tap="{{ onMessageTap }}"/>
-                            </GridLayout>
-                        </Repeater.itemTemplate>
-                    </Repeater>
-                </StackLayout>
-            </ScrollView>
-        </TabViewItem>
-
-        <!-- Profile Tab -->
-        <TabViewItem title="Profile" iconSource="res://profile">
-            <ScrollView>
-                <StackLayout>
-                    <!-- Profile Header -->
-                    <Image src="{{ currentUser.poster_url }}" class="poster-image"/>
-                    <StackLayout class="profile-header">
-                        <Image src="{{ currentUser.dp_url }}" class="profile-image-large"/>
-                        <Label text="{{ currentUser.name }}" class="h1"/>
-                        <Label text="{{ currentUser.bio }}" class="h3"/>
-                    </StackLayout>
-
-                    <!-- Media Grid -->
-                    <GridLayout rows="auto" columns="*,*,*" class="media-grid">
-                        <Repeater items="{{ currentUser.media }}">
-                            <Repeater.itemTemplate>
-                                <Image src="{{ media_url }}" class="media-thumbnail"/>
-                            </Repeater.itemTemplate>
-                        </Repeater>
-                    </GridLayout>
-                </StackLayout>
-            </ScrollView>
-        </TabViewItem>
-    </TabView>
-</Page>
